requirement.txt
# Data visualization
matplotlib
seaborn
# Data manipulation and analysis
pandas
numpy
gymnasium
stable-baselines3


tensorflow
numpy
pandas
tf2onnx
keras
scikit-learn
ta
yfinance
matplotlib
hmmlearn
plotly
QuantStats
ccxt
pandas_ta